requirements.txt
pymodbus==3.7.0
paho-mqtt~=2.1.0
httpx~=0.27.2